feature: Add regex patterns on URL to select which repositories should be cached.
As requested in issue #7 this commit implements the first part of the
regex based mechanism to select which repositories should be cached.
The selection is based on the two patterns GITCACHE_URLPATTERNS_INCLUDE_REGEX
and GITCACHE_URLPATTERNS_EXCLUDE_REGEX (resp. the corresponding entries of the
configuration file). For a repository to be cached, the include pattern must
match and the exclude pattern must not match.